refactor(Cars.java): 이름 비교 부분 수정
메소드 내부의 명확성을 위해서 수정했습니다.
