+++ src/main/java/racingcar/Cars.java
@@ -24,16 +24,12 @@ public class Cars {
         return racingRecords.subList(1, racingRecords.size());
     }
 
-    public Map<String, Boolean> isCarsName(List<String> carNames) {
+    public Map<String, Boolean> matchCarNames(List<String> carNames) {
         Map<String, Boolean> carNamesMatchResult = new HashMap<>();
 
-        for (int i = 0; i < carNames.size(); i++) {
-            if (cars.get(i).isName(carNames.get(i))) {
-                carNamesMatchResult.put(carNames.get(i), true);
-                continue;
-            }
-
-            carNamesMatchResult.put(carNames.get(i), false);
+        int maxSize = Math.max(carNames.size(), this.cars.size());
+        for (int i = 0; i < maxSize; i++) {
+            carNamesMatchResult.put(carNames.get(i), this.cars.get(i).isName(carNames.get(i)));
         }
 
         return carNamesMatchResult;
